requirements.txt
fastapi
uvicorn
openai==0.28
pyyaml